Moved ClimberSubsystem out of "pivot" folder. Added set and stop for climber in command factory. Implemented buttons for climber controls.
